refactor(db): extract taxonomy resolve-or-insert helpers to db/src/taxonomy.rs
